+++ include/pigo/impl/coo.impl.hpp
@@ -1,9 +1,11 @@
 /**
  * PIGO: a parallel graph and matrix I/O and preprocessing library
- * Copyright (c) 2021 GT-TDALab
+ * Copyright (c) 2022 GT-TDALab
  */
 
+#ifdef _OPENMP
 #include <omp.h>
+#endif
 #include <atomic>
 #include <vector>
 #include <type_traits>
@@ -381,8 +383,9 @@ namespace pigo {
     template<class L, class O, class S, bool sym, bool ut, bool sl, bool wgt, class W, class WS>
     void COO<L,O,S,sym,ut,sl,wgt,W,WS>::read_el_(FileReader& r) {
         // Get the number of threads
+        size_t num_threads = 1;
+        #ifdef _OPENMP
         omp_set_dynamic(0);
-        size_t num_threads = 0;
         #pragma omp parallel shared(num_threads)
         {
             #pragma omp single
@@ -390,6 +393,7 @@ namespace pigo {
                 num_threads = omp_get_num_threads();
             }
         }
+        #endif
 
         // This takes two passes:
         // first, count the number of newlines to determine how to
@@ -403,7 +407,11 @@ namespace pigo {
         #pragma omp parallel reduction(max : max_row) \
                 reduction(max : max_col)
         {
+            #ifdef _OPENMP
             size_t tid = omp_get_thread_num();
+            #else
+            size_t tid = 0;
+            #endif
 
             // Find our offsets in the file
             size_t size = r.size();
@@ -567,8 +575,9 @@ namespace pigo {
         // Second, each thread actually writes
 
         // Get the number of threads
+        size_t num_threads = 1;
+        #ifdef _OPENMP
         omp_set_dynamic(0);
-        size_t num_threads = 0;
         #pragma omp parallel shared(num_threads)
         {
             #pragma omp single
@@ -576,12 +585,17 @@ namespace pigo {
                 num_threads = omp_get_num_threads();
             }
         }
+        #endif
 
         std::vector<size_t> pos_offsets(num_threads+1);
         std::shared_ptr<File> f;
         #pragma omp parallel shared(f) shared(pos_offsets)
         {
+            #ifdef _OPENMP
             size_t tid = omp_get_thread_num();
+            #else
+            size_t tid = 0;
+            #endif
             size_t my_size = 0;
 
             #pragma omp for